# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentacion API - Aulas",
        "description": "Listado de endpoints de producto - explicacion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://apiaulas.test/"
        }
    ],
    "paths": {
...
        "/api/entrada_mercancia": {
            "get": {
                "tags": [
                    "Entrada_mercancia"
                ],
                "summary": "Mostrar entradas",
                "operationId": "7987afac908201660778437db8f5e0b1",
                "responses": {
                    "200": {
                        "description": "Mostrar todas las entradas de la tienda.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_entrada": {
                                                        "type": "number",
                                                        "example": "3"
                                                    },
                                                    "cantidad_unidades": {
                                                        "type": "number",
                                                        "example": "20"
                                                    },
                                                    "id_producto": {
                                                        "type": "number",
                                                        "example": "20."
                                                    },
                                                    "id_pedido": {
                                                        "type": "number",
                                                        "example": "2"
                                                    },
                                                    "id_proveedor": {
                                                        "type": "number",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Entrada_mercancia"
                ],
                "summary": "Registrar entrada",
                "operationId": "c433d60c444259d7a3123748a9facddd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cantidad_unidades",
                                    "id_producto",
                                    "id_pedido",
                                    "id_proveedor"
                                ],
                                "properties": {
                                    "cantidad_unidades": {
                                        "type": "number",
                                        "example": 250
                                    },
                                    "id_producto": {
                                        "type": "number",
                                        "example": 1
                                    },
                                    "id_pedido": {
                                        "type": "number",
                                        "example": 2
                                    },
                                    "id_proveedor": {
                                        "type": "number",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pedido guardado exitosamente"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Algo falló al intentar guardar el pedido"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entrada_mercancia/{id_entrada}": {
            "get": {
                "tags": [
                    "Entrada_mercancia"
                ],
                "summary": "Mostrar una entrada específica.",
                "operationId": "6345754de1f53669eae4e9cfa839f5af",
                "parameters": [
                    {
                        "name": "id_entrada",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entrada retornada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_entrada": {
                                            "type": "number",
                                            "example": 3
                                        },
                                        "cantidad_unidades": {
                                            "type": "number",
                                            "example": 20
                                        },
                                        "id_producto": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "id_pedido": {
                                            "type": "number",
                                            "example": 2
                                        },
                                        "id_proveedor": {
                                            "type": "number",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\entrada_mercancia] #id_entrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}